refactor: 구매 횟수 구하는 로직 LottoGame 클래스로 이동
